fix(App): Resolve TS type inference in GA runner loop
The previous loop structure in `runGA` was not correctly inferring the type of `iteration.value` when `iteration.done` was false. This change modifies the loop to use an explicit `else` block, allowing TypeScript to properly narrow `iteration.value` to `ProgressData`, thus resolving type errors.

Also, removed "src" from tsconfig.json `include` as it's redundant when specific directories and files within it are already included.